requirements.txt
streamlit
langchain
openai
tiktoken
python-docx
PyPDF2
pdfplumber
beautifulsoup4
requests
faiss-cpu
